setup ni tax tests

+++ IncomeCalc/src/NITax.hs
@@ -1,4 +1,4 @@
 module NITax (calculateNITax) where
 
-calculateNITax :: Int -> Int
+calculateNITax :: Int -> Double
 calculateNITax x = 0
